fx-azure-test-space/fx-functions/requirements.txt
azure-functions==1.0.0a5
azure-functions-worker==1.0.0a6
html5lib==1.0.1
protobuf==3.6.1
six==1.12.0
webencodings==0.5.1
beautifulsoup4==4.7.1
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
configparser==3.7.1
crayons==0.1.2
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-core==0.29.1
google-cloud-datastore==1.7.3
googleapis-common-protos==1.5.6
grpcio==1.14.2
grpcio-tools==1.14.2
idna==2.8
lxml==4.3.0
html5lib==1.0.1
numpy==1.16.0
pandas==0.24.0
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
soupsieve==1.7.3
urllib3==1.24.1